backend/pyproject.toml
[tool.poetry]
name = "startup"
version = "0.1.0"
description = ""
authors = ["RRoxxxsii <mishabur38@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["all"], version = "^0.109.0"}
sqlalchemy = "^2.0.25"
alembic = "^1.13.1"
uvicorn = "^0.25.0"
pydantic-settings = "^2.1.0"
asyncpg = "^0.29.0"
passlib = "^1.7.4"
psycopg2-binary = "^2.9.9"

[tool.poetry.dev-dependencies]
pytest="^7.4.4"
pre-commit = "^3.6.0"
faker = "^22.7.0"
pytest-asyncio = "^0.23.3"
pytest-mock = "^3.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.pytest.ini_options]
pythonpath = [".", "src"]
python_files = ["test_*.py",]
asyncio_mode="auto"
